Cambios al tipo de dato foreaneas y algunas primarias

+++ CarpetaProyecto/database/migrations/2020_10_14_155548_create_facturas_table.php
@@ -15,7 +15,7 @@ class CreateFacturasTable extends Migration
     {
         Schema::create('factura', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('idUsuario')->unsigned();
+            $table->bigInteger('idUsuario')->unsigned();
             $table->dateTime('fechaCreacion');
             $table->boolean('estado')->default(1);
